yes this is added enemy
