+++ game.py
@@ -11,7 +11,7 @@ def main():
     
 
     # Create goblins ༼ ºل͟º ༽ ༼ ºل͟º ༽ ༼ ºل͟º ༽
-    goblins = [Goblin(f"Goblin {i+1}") for i in range(8)]
+    goblins = [Goblin(f"Goblin {i+1}", "purple") for i in range(8)]
 
     # Keep track of how many goblins were defeated
     defeated_goblins = 0
